KIND Cluster config manifest - my first.
Creates a cluster with the name 'clusterfuck' - because i already know that's what these first few goes are going to be!

It creates 4 Nodes in total, 1 Control-Plane and 3 Workers.

I have assigned Labels:Tier: to the Worker Nodes too, which i can use as selectors later for when it comes to assigning Pods to a Node.

Next I will be creating my deployment manifest